requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Database
sqlite3
pandas==2.0.3

# Document Processing
python-docx==0.8.11
openpyxl==3.1.2

# Desktop Application
webview==4.2.2
pywebview==4.2.2

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Build Tools
pyinstaller==6.3.0

# Optional: Development Tools
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0